requirements.txt
hurry==1.1
hurry.filesize==0.9
numpy==1.23.4
Pillow==9.3.0
streamlit==1.15.2
opencv-python-headless<4.6.0.66

